Refactor: Split PaymentProcessor.tsx into smaller components
Refactors PaymentProcessor.tsx into smaller, more manageable components to improve code organization and readability.
